+++ src/main/kotlin/platform/mixin/inspection/injector/DiscouragedInjectionPointInspection.kt
@@ -1,0 +1,48 @@
+/*
+ * Minecraft Development for IntelliJ
+ *
+ * https://mcdev.io/
+ *
+ * Copyright (C) 2024 minecraft-dev
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published
+ * by the Free Software Foundation, version 3.0 only.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package com.demonwav.mcdev.platform.mixin.inspection.injector
+
+import com.demonwav.mcdev.platform.mixin.handlers.injectionPoint.InjectionPoint
+import com.demonwav.mcdev.platform.mixin.inspection.MixinInspection
+import com.demonwav.mcdev.platform.mixin.util.MixinConstants
+import com.demonwav.mcdev.util.constantStringValue
+import com.intellij.codeInspection.ProblemsHolder
+import com.intellij.psi.JavaElementVisitor
+import com.intellij.psi.PsiAnnotation
+import com.intellij.psi.PsiElementVisitor
+
+class DiscouragedInjectionPointInspection : MixinInspection() {
+    override fun getStaticDescription() = "Reports when a discouraged injection point is used"
+
+    override fun buildVisitor(holder: ProblemsHolder): PsiElementVisitor = object : JavaElementVisitor() {
+        override fun visitAnnotation(annotation: PsiAnnotation) {
+            if (!annotation.hasQualifiedName(MixinConstants.Annotations.AT)) {
+                return
+            }
+            val atValue = annotation.findDeclaredAttributeValue("value") ?: return
+            val atCode = atValue.constantStringValue ?: return
+            val discouragedMessage = InjectionPoint.byAtCode(atCode)?.discouragedMessage
+            if (discouragedMessage != null) {
+                holder.registerProblem(atValue, discouragedMessage)
+            }
+        }
+    }
+}